+++ resources/views/pages/partials/blocks/html.blade.php
@@ -1,4 +1,13 @@
-<div>{!! $block->content !!}</div>
+@php
+    $trustedHtml = app(\App\Support\Blocks\TrustedHtmlOverlayExtractor::class)->extract((string) $block->content);
+    $overlayHtml = $trustedHtml['overlay'] ?? null;
+
+    if ($overlayHtml) {
+        app(\App\Support\Blocks\PublicOverlayRegistry::class)->push($overlayHtml);
+    }
+@endphp
+
+<div>{!! $trustedHtml['content'] ?? $block->content !!}</div>
 
 @if ($block->children->isNotEmpty())
     <div class="wb-stack wb-gap-4">